Evaluator/requirements.txt
ifastapi
uvicorn
python-dotenv
python-multipart
tqdm 

openai